requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
urllib3<2.0.0  # Ensure compatibility with older systems 